+++ src/components/switch/index.tsx
@@ -1,0 +1,32 @@
+"use client"
+
+import * as React from "react"
+import { Switch as SwitchPrimitive } from "radix-ui"
+
+import { cn } from "@/utils/cn"
+
+import styles from "./style.module.sass"
+
+function Switch({
+	className,
+	size = "default",
+	...props
+}: React.ComponentProps<typeof SwitchPrimitive.Root> & {
+	size?: "sm" | "default"
+}) {
+	return (
+		<SwitchPrimitive.Root
+			data-slot="switch"
+			data-size={size}
+			className={cn(styles.root, className)}
+			{...props}
+		>
+			<SwitchPrimitive.Thumb
+				data-slot="switch-thumb"
+				className={styles.thumb}
+			/>
+		</SwitchPrimitive.Root>
+	)
+}
+
+export { Switch }