+++ src/components/table/patientTable/dataTableColumnHeader.tsx
@@ -1,0 +1,67 @@
+import { type Column } from "@tanstack/react-table"
+import { ArrowDown, ArrowUp, ChevronsUpDown, EyeOff } from "lucide-react"
+
+import { cn } from "@/utils/cn"
+import { Button } from "@/components/action/Button"
+import {
+	DropdownMenu,
+	DropdownMenuContent,
+	DropdownMenuItem,
+	DropdownMenuSeparator,
+	DropdownMenuTrigger,
+} from "@/components/dropdown"
+import styles from "./dataTableColumnHeader.module.sass"
+
+interface DataTableColumnHeaderProps<TData, TValue>
+	extends React.HTMLAttributes<HTMLDivElement> {
+	column: Column<TData, TValue>
+	title: string
+}
+
+export function DataTableColumnHeader<TData, TValue>({
+	column,
+	title,
+	className,
+}: DataTableColumnHeaderProps<TData, TValue>) {
+	if (!column.getCanSort()) {
+		return <div className={cn(className)}>{title}</div>
+	}
+
+	return (
+		<div className={cn(styles.headerContainer, className)}>
+			<DropdownMenu>
+				<DropdownMenuTrigger asChild>
+					<Button
+						variant="ghost"
+						size="sm"
+						className={styles.headerButton}
+					>
+						<span>{title}</span>
+						{column.getIsSorted() === "desc" ? (
+							<ArrowDown className={styles.icon} />
+						) : column.getIsSorted() === "asc" ? (
+							<ArrowUp className={styles.icon} />
+						) : (
+							<ChevronsUpDown className={styles.iconUnsorted} />
+						)}
+					</Button>
+				</DropdownMenuTrigger>
+				<DropdownMenuContent align="start">
+					<DropdownMenuItem onClick={() => column.toggleSorting(false)}>
+						<ArrowUp />
+						Asc
+					</DropdownMenuItem>
+					<DropdownMenuItem onClick={() => column.toggleSorting(true)}>
+						<ArrowDown />
+						Desc
+					</DropdownMenuItem>
+					<DropdownMenuSeparator />
+					<DropdownMenuItem onClick={() => column.toggleVisibility(false)}>
+						<EyeOff />
+						Hide
+					</DropdownMenuItem>
+				</DropdownMenuContent>
+			</DropdownMenu>
+		</div>
+	)
+}